+++ src/plugins/require-min-role.ts
