+++ src/routes/tags.ts
@@ -2,8 +2,8 @@ import { FastifyInstance } from 'fastify';
 import { ZodTypeProvider } from 'fastify-type-provider-zod';
 import { EntityNotFoundError } from 'typeorm';
 import { z } from 'zod';
-import { preHandlerBuilder } from '../middlewares/require-min-role';
 import { UserRole } from '../models/User';
+import { preHandlerBuilder } from '../plugins/require-min-role';
 import { ApiTagSchema, TagCreateInputSchema, TagUpdateInputSchema } from '../schemas/Tag';
 import { USER_ID } from '../schemas/User';
 import { tagService } from '../services/TagService';
