+++ src/routes/users.ts
@@ -2,8 +2,8 @@ import { FastifyInstance } from 'fastify';
 import { ZodTypeProvider } from 'fastify-type-provider-zod';
 import { omit } from 'lodash';
 import { z } from 'zod';
-import { preHandlerBuilder } from '../middlewares/require-min-role';
 import { User, UserRole } from '../models/User';
+import { preHandlerBuilder } from '../plugins/require-min-role';
 import { ApiUserSchema, USER_ID, UserCreateInputSchema, UserUpdateInputSchema } from '../schemas/User';
 import { userService } from '../services/UserService';
 import { HttpException, HttpExceptionSchema } from '../utils/HttpException';
